Commented out script that fixes GJS-OSK
GJS-OSK is not yet compatible with GNOME 48, so I had to disable it. Additionally, I had to create a workaround for some files that it wanted to decompress in a location which isn't writeable after build time. I left this script enabled but it is blocking the build process for Fedora 42, so it's gone until GJS-OSK gets updated.